less compiler warnings -> less noise (2x unused variables, 1x comparison between signed and unsigned int)
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13617 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_fluid/fluid_windkessel_optimization.cpp
@@ -320,7 +320,7 @@ void FLD::UTILS::FluidWkOptimizationWrapper::Solve(ParameterList params)
       {
         dserror("[Can't optimise]:Flowrates and pressures have two different dimensions");
       }
-      else if (pressures->size()*optwkmap_.size() != dN_du_->RowDim())
+      else if ((int) (pressures->size()*optwkmap_.size()) != dN_du_->RowDim())
       {
         dserror("[Can't optimise]: Pressures (%d) and state-variables (%d) size don't match",pressures->size(),dN_du_->RowDim()/optwkmap_.size());
       }
